feat: Detach handlers on exit (#12)
Telemetry handlers automatically get detached if they fail to run -
which is fine in general but they do log errors and are noisy and can
cover actual errors.

Since we attach the handlers via a GenServer that we start last, and
therefore stop first, we should explicitly detach the handlers on exit
to ensure we no longer see these errors.